refactor(shared): move internal/metrics → shared/metrics
Deviation from plan (Task 22 planned app/metrics). metrics lives in
shared because shared/github instruments external API calls with
ExternalDuration and ExternalErrorsTotal, and shared/* must not depend
on app/. Task 22 is now a no-op; plan will be updated to reflect.

Also corrects a stale package doc comment claiming workers have no
prometheus dependency — workers do emit metrics through shared/github.
